+++ wxc_sdk/as_api.py
@@ -40,13 +40,13 @@ __all__ = ['AsAccessCodesApi', 'AsAdminAuditEventsApi', 'AsAgentCallerIdApi', 'A
            'AsConferenceControlsApi', 'AsConvergedRecordingsApi', 'AsCustomerExperienceEssentialsApi',
            'AsDECTDevicesApi', 'AsDetailedCDRApi', 'AsDeviceConfigurationsApi', 'AsDeviceSettingsJobsApi',
            'AsDevicesApi', 'AsDialPlanApi', 'AsDigitPatternsApi', 'AsDndApi', 'AsECBNApi', 'AsEventsApi',
-           'AsExecAssistantApi', 'AsFeatureSelector', 'AsForwardingApi', 'AsGroupsApi', 'AsGuestCallingApi',
-           'AsGuestManagementApi', 'AsHotDeskApi', 'AsHotDeskingSigninViaVoicePortalApi', 'AsHotelingApi',
-           'AsHuntGroupApi', 'AsIncomingPermissionsApi', 'AsInternalDialingApi', 'AsJobsApi', 'AsLicensesApi',
-           'AsLocationAccessCodesApi', 'AsLocationEmergencyServicesApi', 'AsLocationInterceptApi', 'AsLocationMoHApi',
-           'AsLocationNumbersApi', 'AsLocationVoicemailSettingsApi', 'AsLocationsApi', 'AsMSTeamsSettingApi',
-           'AsManageNumbersJobsApi', 'AsMePersonalAssistantApi', 'AsMeSettingsApi', 'AsMeetingChatsApi',
-           'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi', 'AsMeetingParticipantsApi',
+           'AsExecAssistantApi', 'AsFeatureAccessApi', 'AsFeatureSelector', 'AsForwardingApi', 'AsGroupsApi',
+           'AsGuestCallingApi', 'AsGuestManagementApi', 'AsHotDeskApi', 'AsHotDeskingSigninViaVoicePortalApi',
+           'AsHotelingApi', 'AsHuntGroupApi', 'AsIncomingPermissionsApi', 'AsInternalDialingApi', 'AsJobsApi',
+           'AsLicensesApi', 'AsLocationAccessCodesApi', 'AsLocationEmergencyServicesApi', 'AsLocationInterceptApi',
+           'AsLocationMoHApi', 'AsLocationNumbersApi', 'AsLocationVoicemailSettingsApi', 'AsLocationsApi',
+           'AsMSTeamsSettingApi', 'AsManageNumbersJobsApi', 'AsMePersonalAssistantApi', 'AsMeSettingsApi',
+           'AsMeetingChatsApi', 'AsMeetingClosedCaptionsApi', 'AsMeetingInviteesApi', 'AsMeetingParticipantsApi',
            'AsMeetingPreferencesApi', 'AsMeetingQandAApi', 'AsMeetingQualitiesApi', 'AsMeetingTranscriptsApi',
            'AsMeetingsApi', 'AsMembershipApi', 'AsMessagesApi', 'AsModeManagementApi', 'AsMonitoringApi',
            'AsMoveUsersJobsApi', 'AsMusicOnHoldApi', 'AsNumbersApi', 'AsOperatingModesApi',
@@ -8636,6 +8636,148 @@ class AsExecAssistantApi(AsPersonSettingsApiChild):
         await self.put(ep, params=params, data=data)
 
 
+class AsFeatureAccessApi(AsApiChild, base='telephony'):
+    """
+    End user Feature Access API
+    """
+    
+    async def read_default(self) -> FeatureAccessSettings:
+        """
+        Read Default Feature Access Settings for Person
+
+        Read the default feature access configuration for users in the organization. It allows administrators to review
+        the baseline feature availability settings that will be applied to new users by default, ensuring consistency
+        in user experience and policy enforcement.
+
+        This API is part of the organizational-level user configuration management for feature access. It is used to
+        define the default settings that control which Webex features are enabled or disabled when users are
+        provisioned. In Control Hub, this corresponds to the "Default User Settings" under Calling or Telephony,
+        providing centralized control over user capabilities across the organization.
+
+        To call this API, an administrator must use a full, or read-only administrator auth token with the
+        `spark-admin:telephony_config_read` scope.
+
+        :rtype: :class:`FeatureAccessSettings`
+        """
+        url = self.ep('config/people/settings/permissions')
+        data = await super().get(url)
+        r = FeatureAccessSettings.model_validate(data)
+        return r
+
+    async def update_default(self, settings: FeatureAccessSettings):
+        """
+        Update Default Person Feature Access Configuration
+
+        Updates the default feature access configuration for users in the organization. It allows administrators to
+        modify the baseline settings that determine which Webex features are enabled or disabled for users by default,
+        ensuring new users are provisioned with consistent access controls.
+
+        This API is part of the organization-level user configuration management for feature access. It supports
+        defining and updating default settings that apply automatically to all newly onboarded users. In Control Hub,
+        this corresponds to the "Default User Settings" section for Calling or Telephony, enabling centralized and
+        scalable configuration of user capabilities.
+
+        To use this API, an administrator must authenticate with a full, or device administrator auth token. The token
+        must include the `spark-admin:telephony_config_write` scope.
+
+        :param settings: The feature access settings to be updated.
+        :type settings: :class:`FeatureAccessSettings`
+
+        :rtype: None
+        """
+        body = settings.update()
+        url = self.ep('config/people/settings/permissions')
+        await super().put(url, json=body)
+
+    async def read(self, person_id: str) -> UserFeatureAccessSettings:
+        """
+        Read Feature Access Settings for a Person
+
+        Read the feature access configuration for the current user within the organization. It allows administrators to
+        read the telephony settings, including device and location configurations, specific to that user’s role and
+        access privileges. This API is useful for managing and verifying user-specific feature access within the
+        broader telephony system.
+
+        The feature is part of the organization’s telephony configuration management. It provides insight into the
+        settings and permissions that control how telephony services are assigned and configured for individual users.
+        This functionality is available through the Control Hub and allows for the management of user access to
+        various telephony-related features.
+
+        To access this API, the user must possess a full, or read-only administrator role. The authentication token
+        used must include the `spark-admin:telephony_config_read` scope, granting the necessary permissions to read
+        the telephony configuration for the user in question.
+
+        :param person_id: User ID of the Organization.
+        :type person_id: str
+        :rtype: :class:`UserFeatureAccessSettings`
+        """
+        url = self.ep(f'config/people/{person_id}/settings/permissions')
+        data = await super().get(url)
+        r = UserFeatureAccessSettings.model_validate(data)
+        return r
+
+    async def update(self,
+               person_id: str,
+               settings: FeatureAccessSettings):
+        """
+        Update a Person’s Feature Access Configuration
+
+        Update the feature access configuration for the current user within the organization. It enables administrators
+        to modify the telephony settings, including device and location configurations, specific to the user’s role
+        and access privileges. This API is useful for making adjustments to user-specific feature access within the
+        telephony system.
+
+        The feature is part of the organization’s telephony configuration management. It provides control over the
+        settings and permissions that govern how telephony services are assigned and configured for individual users.
+        This functionality is available through the Control Hub and enables the modification of user access to various
+        telephony-related features.
+
+        To use this API, an administrator must authenticate with a full, or device administrator auth token. The
+        authentication token used must include the `spark-admin:telephony_config_write` scope, granting the necessary
+        permissions to update the telephony configuration for the user in question.
+
+
+        :param person_id: User ID of the Organization.
+        :type person_id: str
+        :param settings: The feature access settings to be updated.
+        :type settings: :class:`FeatureAccessSettings`
+        :rtype: None
+        """
+        body = settings.update()
+        url = self.ep(f'config/people/{person_id}/settings/permissions')
+        await super().put(url, json=body)
+
+    async def reset(self, person_id: str):
+        """
+        Reset a Person’s Feature Access Configuration to the Organization’s Default Settings
+
+        Reset of a user’s feature access configuration to the organization’s default settings. It ensures that any
+        specific feature configurations set by an administrator for an individual user are overridden and replaced
+        with the global configuration of the organization. This process helps to maintain consistency in feature
+        access across all users, especially when administrators want to ensure that a user is subject to the
+        organization's global settings rather than personalized settings.
+
+        The overall feature, managed through the organization's Control Hub, involves the configuration and
+        customization of feature access for users. Administrators can tailor these settings to individual users based
+        on their roles or needs, but sometimes a global reset to the default configuration is necessary. The reset API
+        simplifies this by programmatically resetting a user’s feature access, which can be crucial when managing
+        large teams or updating organizational policies that affect user privileges across multiple devices or
+        locations.
+
+        To use this API, an administrator must authenticate with a full, or device administrator auth token. This
+        ensures the individual has the necessary privileges to make changes to user configurations. Furthermore, the
+        authentication token used must include the `spark-admin:telephony_config_write` scope, which grants the
+        required permissions to modify the telephony configuration for the user. This combination of roles and scopes
+        ensures that only authorized administrators can reset the feature access configuration.
+
+        :param person_id: User ID of the Organization.
+        :type person_id: str
+        :rtype: None
+        """
+        url = self.ep(f'config/people/{person_id}/settings/permissions/actions/reset/invoke')
+        await super().post(url)
+
+
 class AsHotelingApi(AsPersonSettingsApiChild):
     """
     API for person's hoteling settings
@@ -11367,6 +11509,8 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
     ecbn: AsECBNApi
     #: Executive Assistant Settings for a Person
     exec_assistant: AsExecAssistantApi
+    #: Featuer access settings for a person
+    feature_access: AsFeatureAccessApi
     #: Forwarding Settings for a Person
     forwarding: AsPersonForwardingApi
     #: Hoteling Settings for a Person
@@ -11429,6 +11573,7 @@ class AsPersonSettingsApi(AsApiChild, base='people'):
         self.dnd = AsDndApi(session=session)
         self.ecbn = AsECBNApi(session=session)
         self.exec_assistant = AsExecAssistantApi(session=session)
+        self.feature_access = AsFeatureAccessApi(session=session)
         self.forwarding = AsPersonForwardingApi(session=session)
         self.hoteling = AsHotelingApi(session=session)
         self.mode_management = AsModeManagementApi(session=session)